--- webshopHomework/mobilewebshop/mobilewebshop-ejb/src/main/java/com/mycompany/ee/constraint/BirthDayInPastValidator.java
@@ -1,21 +1,0 @@
-package com.mycompany.ee.constraint;
-
-import java.time.LocalDate;
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-
-public class BirthDayInPastValidator implements ConstraintValidator<BirthDayInPastConstraint, LocalDate>{
-
-    @Override
-    public void initialize(BirthDayInPastConstraint constraintAnnotation) {
-    }
-    
-   @Override
-   public boolean isValid(LocalDate date, ConstraintValidatorContext context) {
-       if(date!= null){
-           return date.isBefore(LocalDate.now());
-       }
-       return true;
-   }
-}
