--- webshopHomework/mobilewebshop/mobilewebshop-ejb/src/main/java/com/mycompany/ee/producer/LoggingProducer.java
@@ -1,14 +1,0 @@
-package com.mycompany.ee.producer;
-
-import java.util.logging.Logger;
-import javax.enterprise.inject.Produces;
-import javax.enterprise.inject.spi.InjectionPoint;
-
-
-public class LoggingProducer {
-
-    @Produces
-    public Logger produceLogger(InjectionPoint injectionPoint) {
-        return Logger.getLogger(injectionPoint.getMember().getDeclaringClass().getName());
-    }
-}
