--- webshopHomework/mobilewebshop/mobilewebshop-ejb/src/main/java/com/mycompany/ee/producer/ValidatorProducer.java
@@ -1,16 +1,0 @@
-package com.mycompany.ee.producer;
-
-import javax.validation.Validation;
-import javax.validation.Validator;
-import javax.validation.ValidatorFactory;
-import javax.ws.rs.Produces;
-
-
-public class ValidatorProducer {
-
-    @Produces
-    public Validator produceLogger() {
-        ValidatorFactory vf = Validation.buildDefaultValidatorFactory();
-        return vf.getValidator();
-    }
-}
