--- webshopHomework/mobilewebshop/mobilewebshop-web/src/main/java/com/mycompany/ee/exceptionmapper/ErrorDTO.java
@@ -1,22 +1,0 @@
-package com.mycompany.ee.exceptionmapper;
-
-
-public class ErrorDTO {
-
-    private String errorMessage;
-
-    public ErrorDTO() {
-    }
-
-    public ErrorDTO(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-
-    public void setErrorMessage(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }   
-}
