+++ webshopHomework/sza-mobilewebshop/pom.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>com.mycompany</groupId>
+  <artifactId>sza-mobilewebshop</artifactId>
+  <version>1.0-SNAPSHOT</version>
+  <packaging>pom</packaging>
+  <name>sza-mobilewebshop</name>
+  <modules>
+    <module>sza-mobilewebshop-ear</module>
+    <module>sza-mobilewebshop-web</module>
+    <module>sza-mobilewebshop-ejb</module>
+  </modules>
+</project>
