+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ear/nb-configuration.xml
