+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ear/src/main/application/META-INF/MANIFEST.MF
