+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ear/src/main/application/META-INF/jboss-app.xml
