+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ejb/nb-configuration.xml
