+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ejb/src/main/java/com/mycompany/ee/exception/BadRequestException.java
