+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ejb/src/main/java/com/mycompany/ee/producer/LoggingProducer.java
@@ -1,0 +1,11 @@
+package com.mycompany.ee.producer;
+
+import java.util.logging.Logger;
+import javax.enterprise.inject.spi.InjectionPoint;
+
+public class LoggingProducer {
+
+    public Logger produceLogger(InjectionPoint injectionPoint) {
+        return Logger.getLogger(injectionPoint.getMember().getDeclaringClass().getName());
+    }
+}
