+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ejb/src/main/java/com/mycompany/ee/producer/ValidatorProducer.java
@@ -1,0 +1,13 @@
+package com.mycompany.ee.producer;
+
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
+
+public class ValidatorProducer {
+
+    public Validator produceLogger() {
+        ValidatorFactory vf = Validation.buildDefaultValidatorFactory();
+        return vf.getValidator();
+    }
+}
