+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-ejb/src/main/resources/META-INF/MANIFEST.MF
