+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/nb-configuration.xml
