+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/java/com/mycompany/ee/exceptionmapper/BadRequestExceptionMapper.java
