+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/java/com/mycompany/ee/exceptionmapper/ErrorDTO.java
@@ -1,0 +1,23 @@
+package com.mycompany.ee.exceptionmapper;
+
+
+public class ErrorDTO {
+
+    private String errorMessage;
+
+    public ErrorDTO() {
+        //empty constructor
+    }
+
+    public ErrorDTO(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+
+    public void setErrorMessage(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }   
+}
