+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/java/com/mycompany/rest/InventoryRESTService.java
