+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/java/org/netbeans/rest/application/config/ApplicationConfig.java
