+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/resources/META-INF/jboss.xml
