+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/webapp/WEB-INF/beans.xml
