+++ webshopHomework/sza-mobilewebshop/sza-mobilewebshop-web/src/main/webapp/index.html
